+++ internal/engine/ingester/diff/diff.go
@@ -215,8 +215,9 @@ func ingestFileForFullDiff(filename, patch, patchUrl string) (*pb.PrContents_Fil
 			}
 		} else if strings.HasPrefix(line, "+") {
 			result = append(result, &pb.PrContents_File_Line{
-				Content:    line[1:],
-				LineNumber: currentLineNumber,
+				Content: line[1:],
+				// see the use of strconv.ParseInt above: this is a safe downcast
+				LineNumber: int32(currentLineNumber),
 			})
 
 			currentLineNumber++
